steps 2-4
