+++ packages/manager/src/features/PlacementGroups/PlacementGroupsLanding/PlacementGroupsLanding.tsx
@@ -1,6 +1,5 @@
 import { CircleProgress } from '@linode/ui';
 import { useMediaQuery, useTheme } from '@mui/material';
-import { createLazyRoute } from '@tanstack/react-router';
 import * as React from 'react';
 import { useParams } from 'react-router-dom';
 import { useHistory } from 'react-router-dom';
@@ -280,10 +279,4 @@ export const PlacementGroupsLanding = React.memo(() => {
       />
     </>
   );
-});
-
-export const placementGroupsLandingLazyRoute = createLazyRoute(
-  '/placement-groups'
-)({
-  component: PlacementGroupsLanding,
 });
