+++ packages/manager/src/routes/index.tsx
@@ -87,6 +87,7 @@ declare module '@tanstack/react-router' {
 export const migrationRouteTree = migrationRootRoute.addChildren([
   betaRouteTree,
   domainsRouteTree,
+  placementGroupsRouteTree,
   longviewRouteTree,
   volumesRouteTree,
 ]);
