+++ packages/manager/src/routes/placementGroups/placementGroupsLazyRoutes.ts
@@ -1,0 +1,22 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { PlacementGroupsDetail } from 'src/features/PlacementGroups/PlacementGroupsDetail/PlacementGroupsDetail';
+import { PlacementGroupsLanding } from 'src/features/PlacementGroups/PlacementGroupsLanding/PlacementGroupsLanding';
+
+export const placementGroupsDetailLazyRoute = createLazyRoute(
+  '/placement-groups/$id'
+)({
+  component: PlacementGroupsDetail,
+});
+
+export const placementGroupsUnassignLazyRoute = createLazyRoute(
+  '/placement-groups/$id/linodes/unassign/$linodeId'
+)({
+  component: PlacementGroupsDetail,
+});
+
+export const placementGroupsLandingLazyRoute = createLazyRoute(
+  '/placement-groups'
+)({
+  component: PlacementGroupsLanding,
+});
